Revision of how CD/DVDs are accessed to avoid reported issues.
Because of reported issues with accessing the CD in the /target environment
if it is already mounted in the D-I environment, base-installer has been
reverted to loop-mounting the installation CD. In order to still allow
CD-changing while avoiding these issues and keeping installs idempotent, we
now need to allow apt-cdrom to manage mounting/unmounting for all types of
CD images (except for hd-media installs). See: #474346, #475639.

Requires: base-installer (>= 1.90), pkgsel (>= 0.19), di-utils (>= 1.57).

r52884