Revert extra mounts to support libudev for apt-cdrom
Reason for the revert is that the changes are not needed now that APT
has been reconfigured to use media/cdrom in target. The change was also
incomplete as apt-cdrom is also called from other places than 40cdrom.

It may be worth reintroducing this change later, but I prefer to implement
and test the switch from /cdrom to /media/cdrom without it first.

r62633