improve prompting for mirror when no network is configured
In the case of installing from a full CD without a network, display the
mirror use question at high priority, defaulting to no. This lets the user
see explicitly that no mirror will be used (or try to use one and get
the obvious failure), as well as possibly notice the innocuous warning
message about this not being the ideal case, though one that should work.

In the case of installing from a netinst without a network, instead show
the more intrusive apt-setup/no_mirror.