Add support for backports. Closes: #691651
The text added in apt-setup/services-select was adapted from
apt-setup/backports for Ubuntu.

Lunar's patch failed to apply, so hand-applied it.

Added a bit more explanation about the default apt pinning that prevents
backports from being installed by default. Did not mention pinning (jargon)
or apt-get -t $release-backports install ($release not trivially available
and the user should not need to take notes of commands during the install.

Note that the apt line for backports changed with wheezy. If a
squeeze-support udeb is added, this will need to be modified to generate
the right line for squeeze. There is currently no such udeb, so I didn't
bother.

The code duplication between 93backports and 92updates should probably be
eliminated by refactoring them to eg, a shared shell library.